pyproject.toml
[project]
name = "invoice-ocr-demo"
version = "1.0.0"
description = "Invoice OCR Demo Application using Gemini 2.0 Flash"
authors = [
    {name = "Moeed", email = "contact@meetmoeed.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "streamlit>=1.45.0",
    "google-generativeai>=0.8.0",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "pdf2image>=1.17.0",
    "python-dateutil>=2.9.0",
    "openpyxl>=3.1.0",
    "streamlit-lottie>=0.0.5",
    "streamlit-aggrid>=1.0.0",
    "requests>=2.32.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0"
]
